Moved some stuff to shared, added language check for rev conversion
